+++ src/main/java/net/danh/clientcore/block/BlockFarmingStage.java
@@ -1,0 +1,12 @@
+package net.danh.clientcore.block;
+
+import org.bukkit.configuration.ConfigurationSection;
+
+import java.util.List;
+
+record BlockFarmingStage(
+        String block,
+        int afterTicks,
+        List<ConfigurationSection> drops
+) {
+}